refactor: remove private re-exports from ipeps.py, update test imports
Update all test files to import directly from ipeps submodules
(ipeps_config, ipeps_ctm, ipeps_rdm, ipeps_simple_update,
ipeps_optimize) instead of through ipeps.py. Remove 44 private
symbol re-exports from ipeps.py, eliminating the noqa: F401
compatibility façade.

ipeps.py now only imports what it uses locally (370 lines, down
from 423).
